feat: update readme & rename file
